+++ ຂໍ້ມູນຫ້ອງ.cs
@@ -18,9 +18,13 @@ namespace C_Sharp_2_Project
         SqlConnection con = Connect.getconnect();
         SqlCommand cmd;
         SqlDataAdapter da;
+        
         SqlDataReader reader;
         DataSet ds = new DataSet();
+        DataSet ds1 = new DataSet();
+       
         string sql = "";
+
         public ຂໍ້ມູນຫ້ອງ()
         {
             InitializeComponent();
@@ -30,130 +34,193 @@ namespace C_Sharp_2_Project
         {
             showdata();
             comtype();
-           
-            comfloor();
         }
-        string[] strtitle = { "ລະຫັດ", "ເບີຫ້ອງ", "ຊື່ປະເພດ", "ເບີຊັ້ນ", "ຈຳນວນນັກຮຽນ" };
-        // int[] siz = { 50, 50, 80, 50, 50, 100, 130 };
-       // AutoCompleteStringCollection auto = new AutoCompleteStringCollection();
+
+
         private void showdata()
         {
             try
             {
-                sql = "select roomID,room_number,typename,floornumber,Student_amount from tbroom left join tbroomtype on tbroomtype.roomtypeID=tbroom.roomtypeID left join tbfloor on tbfloor.floorID=tbroom.floorID";
-            da = new SqlDataAdapter(sql, con);
-            da.Fill(ds, "room");
-            if (ds.Tables["room"] !=null)
-            {
-                ds.Tables.Clear();
-            }da.Fill(ds, "room");
-            dgvroom.DataSource = ds.Tables["room"];
-            for (int i=0;i<strtitle.Length; i++)
-            {
-                dgvroom.Columns[i].HeaderText = strtitle[i];
-               // dgvroom.Columns[i].Width = siz[i];
+                sql = "select roomID,room_number,typename,floornumber,buildingnumber from tbroom left join tbroomtype on tbroomtype.roomtypeID=tbroom.roomtypeID left join tbfloor on tbfloor.floorID=tbroom.floorID inner join tbbuilding on tbfloor.buildingID= tbbuilding.buildingID";
+                da = new SqlDataAdapter(sql, con);
+                da.Fill(ds, "room");
+                if (ds.Tables["room"] != null)
+                {
+                    ds.Tables.Clear();
+                }
+                da.Fill(ds, "room");
+                dgvroom.DataSource = ds.Tables["room"];
+                dgvroom.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
+                dgvroom.Columns[0].HeaderText = "ລະຫັດ";
+                dgvroom.Columns[1].HeaderText = "ເບີຫ້ອງ";
+                dgvroom.Columns[2].HeaderText = "ປະເພດ";
+                dgvroom.Columns[3].HeaderText = "ຊັ້ນ";
+                dgvroom.Columns[4].HeaderText = "ຕຶກ";
+
             }
-            dgvroom.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
-            }
-            catch (Exception ex )
+            catch (Exception ex)
             {
 
                 MessageBox.Show(ex.Message.ToString());
-            }        }
+            }
+        }
+        private void builcombo()
+        {
+            sql = "select buildingnumber from tbbuilding inner join tbroomtype on tbroomtype.roomtypeID=tbbuilding.roomtypeID where typename= N'" + cbroomtype.SelectedItem + "'";
+            cmd = new SqlCommand(sql, con);
+            reader = cmd.ExecuteReader();
+            cbbuild.Items.Clear();
+            while (reader.Read())
+            {
+                cbbuild.Items.Add(reader.GetString(0));
+            }
+            reader.Close();
 
+        }
         public void comtype()
         {
             try
             {
                 sql = "select * from tbroomtype";
-             cmd=new SqlCommand (sql, con);
+                cmd = new SqlCommand(sql, con);
                 reader = cmd.ExecuteReader();
                 cbroomtype.Items.Clear();
-                while (reader .Read())
+                while (reader.Read())
                 {
+
                     string typename = reader.GetString(1);
                     cbroomtype.Items.Add(typename);
-                }reader.Close();
-                
+                }
+                reader.Close();
+
             }
             catch (Exception ex)
             {
                 MessageBox.Show(ex.Message.ToString());
-        }
+            }
         }
 
-      void clrtext()
-        {
+        void clrtext()
+        {   
+          cbroomtype.SelectedIndex=-1;
             txtroomid.Clear();
             txtroomnum.Clear();
-            cbfloor.SelectedIndex =-1;
-            cbfloor. ResetText ();
-            cbroomtype.SelectedIndex = -1;
-            cbroomtype.ResetText();
-            txtstamount.Clear();
-            txtroomid.Focus();
+            cbfloor.SelectedIndex = -1;          
+            cbbuild.SelectedIndex=-1;
+            
+            
         }
-           
-       
+
+
         public void comfloor()
         {
             sql = "select distinct floornumber from tbfloor ";
             cmd = new SqlCommand(sql, con);
             reader = cmd.ExecuteReader();
             cbfloor.Items.Clear();
-            while (reader .Read())
+            while (reader.Read())
             {
                 string strflo = reader.GetString(0);
                 cbfloor.Items.Add(strflo);
-            }reader.Close();
+            }
+            reader.Close();
         }
 
 
         private void btrsave_Click(object sender, EventArgs e)
         {
             try
-            { 
-                sql = "insert into tbroom values(@roomID,@roomnum,@rtype,(select floorID from tbfloor where floornumber=@flo),@stamount)";
-            cmd = new SqlCommand(sql, con);
-            cmd.Parameters.Clear();
-            cmd.Parameters.AddWithValue("@roomID",txtroomid .Text );
-            cmd.Parameters.AddWithValue("@roomnum",txtroomnum.Text);
-            cmd.Parameters.AddWithValue("@rtype",cbroomtype .SelectedIndex+1);
-            cmd.Parameters.AddWithValue("@flo",cbfloor .SelectedItem);
-            //cmd.Parameters.AddWithValue("@bui",cbbuildID .SelectedIndex +1);         
-            cmd.Parameters.AddWithValue("@stamount",txtstamount .Text );
-            cmd.ExecuteNonQuery();
-            showdata();
-                comtype();               
-                comfloor();
-                clrtext();
+            {
+                if (txtroomid .Text ==null || txtroomnum.Text ==null || cbbuild.Text ==null || cbfloor .Text ==null || cbroomtype .Text ==null)
+                {
+                    MessageBox.Show("ຂໍ້ມູນບໍ່ຄົບ.");
+                    return;
+                }
+                string s = "select room_number from tbroom where room_number ='" + txtroomnum.Text +"'";
+                cmd = new SqlCommand(s, con);
+                da = new SqlDataAdapter(cmd);
+                DataTable dt = new DataTable();
+                da.Fill(dt);
+                int i = dt.Rows.Count;
+                if (i >=1)
+                {
+                    MessageBox.Show("ເບີຫ້ອງນີ້ມີແລ້ວ.", "Duplicate", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                  
+                }
 
+                /* string s = "select room_number from tbroom";
+                  SqlCommand cmd1 = new SqlCommand(s, con);
+                //cmd.Parameters.AddWithValue("@nn", txtroomnum.Text);
+                  SqlDataReader rer = cmd.ExecuteReader();
+                     while (rer.Read())
+                  {
+                      if (rer[0].ToString()==txtroomnum .Text )
+                     {
+                        MessageBox.Show("ເບີຫ້ອງນີ້ມີແລ້ວ.", "Duplicate", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                     }
+                  }
+                  rer.Close();*/
+
+                //#pragma warning disable CS0162 // Unreachable code detected
+                /* int o = dgvroom.Rows.Count;
+ #pragma warning disable CS0162 // Unreachable code detected
+                 for (int i = 0; i < o; i++)
+                 {
+
+                     string dd = dgvroom.Rows[i].Cells[1].Value.ToString();
+                     if (txtroomnum.Text == dd)
+                     {
+                         MessageBox.Show("ເບີຫ້ອງນີ້ມີແລ້ວ.", "Duplicate", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                         return;
+                     }*/
+                else
+                {
+                    sql = @"insert into tbroom values(@roomnum,@rtype,(select floorID from tbfloor where floornumber=@flo),@stm)";
+                    cmd = new SqlCommand(sql, con);
+                    cmd.Parameters.Clear();
+                    // cmd.Parameters.AddWithValue("@roomID",txtroomid .Text );
+                    cmd.Parameters.AddWithValue("@roomnum", txtroomnum.Text);
+                    cmd.Parameters.AddWithValue("@rtype", cbroomtype.SelectedIndex + 1);
+                    cmd.Parameters.AddWithValue("@flo", cbfloor.SelectedItem);
+                    cmd.Parameters.AddWithValue("@stm", 0);
+                    cmd.ExecuteNonQuery();
+                    showdata();                   
+                    clrtext();
+
+                }
+#pragma warning restore CS0162 // Unreachable code detected
+                //}
             }
             catch (Exception ex)
             {
 
                 MessageBox.Show(ex.Message.ToString());
             }
-           
+
         }
 
         private void btredit_Click(object sender, EventArgs e)
         {
             try
             {
-                sql = "update tbroom set room_number=@rnum,roomtypeID=@type,floorID=(select floorID from tbfloor where floornumber=@floid),Student_amount=@stamount where roomID=@rID";
+                if (txtroomid.Text == null || txtroomnum.Text == null || cbbuild.Text == null || cbfloor.Text == null || cbroomtype.Text == null)
+                {
+                    MessageBox.Show("ຂໍ້ມູນບໍ່ຄົບ.");
+                    return;
+                }
+                sql = "update tbroom set room_number=@rnum,roomtypeID=@type,floorID=(select floorID from tbfloor where floornumber=@floid) where roomID=@rID";
                 cmd = new SqlCommand(sql, con);
                 cmd.Parameters.Clear();
-                cmd.Parameters.AddWithValue("@rnum",txtroomnum .Text);
-                cmd.Parameters.AddWithValue("@type",cbroomtype .SelectedIndex+1);
-                cmd.Parameters.AddWithValue("@floid",cbfloor .SelectedItem);
-              //  cmd.Parameters.AddWithValue("@buid",cbbuildID .SelectedIndex+1);
-               
-                cmd.Parameters.AddWithValue("@stamount",txtstamount .Text );
-                cmd.Parameters.AddWithValue("@rID",txtroomid .Text );
+                cmd.Parameters.AddWithValue("@rnum", txtroomnum.Text);
+                cmd.Parameters.AddWithValue("@type", cbroomtype.SelectedIndex + 1);
+                cmd.Parameters.AddWithValue("@floid", cbfloor.SelectedItem);
+                //  cmd.Parameters.AddWithValue("@buid",cbbuildID .SelectedIndex+1);               
+                // cmd.Parameters.AddWithValue("@stamount",txtstamount .Text );
+                cmd.Parameters.AddWithValue("@rID", txtroomid.Text);
                 cmd.ExecuteNonQuery();
-                showdata();             
+                showdata();
                 comtype();
+                builcombo();
                 comfloor();
                 clrtext();
 
@@ -169,6 +236,11 @@ namespace C_Sharp_2_Project
         {
             try
             {
+                if (txtroomid.Text == null)
+                {
+                    MessageBox.Show("ຂໍ້ມູນບໍ່ຄົບ.");
+                    return;
+                }
                 sql = "delete from tbroom where roomID=@roomID";
                 cmd = new SqlCommand(sql, con);
                 cmd.Parameters.Clear();
@@ -176,7 +248,6 @@ namespace C_Sharp_2_Project
                 cmd.ExecuteNonQuery();
                 showdata();
                 comtype();
-               
                 comfloor();
                 clrtext();
 
@@ -191,19 +262,18 @@ namespace C_Sharp_2_Project
 
         private void dgvroom_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
-            txtroomid.Text = dgvroom.CurrentRow.Cells[0].Value.ToString();
-             txtroomnum.Text = dgvroom.CurrentRow.Cells[1].Value.ToString();
-            cbroomtype.SelectedItem  = dgvroom.CurrentRow.Cells[2].Value.ToString ();
-            cbfloor.SelectedItem = dgvroom.CurrentRow.Cells[3].Value.ToString ();
-
-            txtstamount.Text = dgvroom.CurrentRow.Cells[4].Value.ToString();
+            txtroomid.Text = dgvroom.CurrentRow.Cells[0].Value.ToString(); 
+            cbroomtype.SelectedItem = dgvroom.CurrentRow.Cells[2].Value.ToString();
+            cbbuild.SelectedItem = dgvroom.CurrentRow.Cells[4].Value.ToString();
+            cbfloor.SelectedItem = dgvroom.CurrentRow.Cells[3].Value.ToString();
+            txtroomnum.Text = dgvroom.CurrentRow.Cells[1].Value.ToString();
         }
 
         private void txtsearch_TextChanged(object sender, EventArgs e)
         {
             try
             {
-                sql = "select roomID,room_number,typename,floornumber,Student_amount from tbroom left join tbroomtype on tbroomtype.roomtypeID=tbroom.roomtypeID left join tbfloor on tbfloor.floorID=tbroom.floorID where roomID='"+ txtsearch.Text +"'";
+                sql = "select roomID,room_number,typename,floornumber from tbroom left join tbroomtype on tbroomtype.roomtypeID=tbroom.roomtypeID left join tbfloor on tbfloor.floorID=tbroom.floorID where roomID='" + txtsearch.Text + "' or room_number='" + txtsearch.Text + "'";
                 da = new SqlDataAdapter(sql, con);
                 da.Fill(ds, "room");
                 if (ds.Tables["room"] != null)
@@ -212,11 +282,12 @@ namespace C_Sharp_2_Project
                 }
                 da.Fill(ds, "room");
                 dgvroom.DataSource = ds.Tables["room"];
-                for (int i = 0; i < strtitle.Length; i++)
-                {
-                    dgvroom.Columns[i].HeaderText = strtitle[i];
-                    // dgvroom.Columns[i].Width = siz[i];
-                }
+                dgvroom.Columns[0].HeaderText = "ລະຫັດ";
+                dgvroom.Columns[1].HeaderText = "ເບີຫ້ອງ";
+                dgvroom.Columns[2].HeaderText = "ປະເພດ";
+                dgvroom.Columns[3].HeaderText = "ຊັ້ນ";
+                dgvroom.Columns[4].HeaderText = "ຕຶກ";
+
                 dgvroom.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
 
             }
@@ -233,15 +304,32 @@ namespace C_Sharp_2_Project
         {
             try
             {
-                sql = "select roomID,room_number,typename,floornumber,room_status from tbroom left join tbroomtype on tbroomtype.roomtypeID=tbroom.roomtypeID left join tbfloor on tbfloor.floorID=tbroom.floorID where roomID='" + txtsearch.Text + "'";
-                da = new SqlDataAdapter(sql, con);
-                da.Fill(ds, "room");
-                if (ds.Tables["room"] != null)
+                if (txtsearch.Text == "")
                 {
-                    ds.Tables.Clear();
+                    showdata();
                 }
-                da.Fill(ds, "room");
+                else
+                {
+                    sql = "select roomID,room_number,typename,floornumber from tbroom left join tbroomtype on tbroomtype.roomtypeID=tbroom.roomtypeID left join tbfloor on tbfloor.floorID=tbroom.floorID where roomID='" + txtsearch.Text + "'";
+                    da = new SqlDataAdapter(sql, con);
+                    da.Fill(ds, "room");
+                    if (ds.Tables["room"] != null)
+                    {
+                        ds.Tables.Clear();
+                    }
+                    da.Fill(ds, "room");
+                    dgvroom.DataSource = ds.Tables["room"];
+
+                }
                 dgvroom.DataSource = ds.Tables["room"];
+                dgvroom.Columns[0].HeaderText = "ລະຫັດ";
+                dgvroom.Columns[1].HeaderText = "ເບີຫ້ອງ";
+                dgvroom.Columns[2].HeaderText = "ປະເພດ";
+                dgvroom.Columns[3].HeaderText = "ຊັ້ນ";
+                dgvroom.Columns[4].HeaderText = "ຕຶກ";
+
+                dgvroom.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
+
 
             }
             catch (Exception ex)
@@ -253,15 +341,10 @@ namespace C_Sharp_2_Project
 
         private void cbroomtype_SelectedIndexChanged(object sender, EventArgs e)
         {
-            sql = "select floornumber from tbfloor where buildingID in (select buildingID from tbbuilding inner join tbroomtype on tbbuilding.roomtypeID=tbroomtype.roomtypeID where typename=N'" + cbroomtype .SelectedItem +"')";
-            cmd = new SqlCommand(sql, con);
-            cbfloor.Items.Clear();
-            reader = cmd.ExecuteReader();
-                while(reader .Read())
-            {
-                cbfloor.Items.Add(reader.GetString(0));
-            }
-            reader.Close();
+            cbbuild.ResetText();
+            cbfloor.ResetText();
+            builcombo();
+
         }
 
         private void cbfloor_SelectedIndexChanged(object sender, EventArgs e)
@@ -279,6 +362,35 @@ namespace C_Sharp_2_Project
         private void label1_Click(object sender, EventArgs e)
         {
             this.Close();
+        }
+
+        private void label10_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void cbbuild_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            cbfloor.Items.Clear();
+            sql = "select floornumber from tbfloor where buildingID in (select buildingID from tbbuilding  where buildingnumber=N'" + cbbuild.SelectedItem + "')";
+            cmd = new SqlCommand(sql, con);
+            cbfloor.Items.Clear();
+            reader = cmd.ExecuteReader();
+            while (reader.Read())
+            {
+                cbfloor.Items.Add(reader.GetString(0));
+            }
+            reader.Close();
+        }
+
+        private void txtroomnum_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void groupBox1_Enter(object sender, EventArgs e)
+        {
+
         }
     }
 }
